Move CentOS family to DistributionInstaller
