+++ mkosi/distributions/alma.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: LGPL-2.1+
+
+from mkosi.distributions.centos import CentosInstaller
+
+
+class AlmaInstaller(CentosInstaller):
+    pass
