+++ mkosi/distributions/alma_epel.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: LGPL-2.1+
+
+from mkosi.distributions.alma import AlmaInstaller
+
+
+class AlmaEpelInstaller(AlmaInstaller):
+    pass
